backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow>=10.2.0
django-storages==1.14.2
boto3==1.34.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.3
PyJWT==2.8.0
xlsxwriter==3.2.0
